refactor(frontend): separate state interfaces and state object from the reducers file
